feat: DishViewModel 연동
- onApear에서 ViewModel이 DishPublisher를 subcribe하도록 설정
- ViewModel의 state에 따라 body에 표시되는 View가 바뀜.
- 오류가 난 경우 다시 시도할 수 있도록 reload버튼 생성
- 디렉토리 세분화
