+++ SideDish/SideDish/Common/Extension/Color.swift
@@ -10,4 +10,5 @@ import SwiftUI
 extension Color {
     static let themeColor: Color = Color("ThemeColor")
     static let backgroundColor: Color = Color("Background")
+    static let labelColor: Color = Color("Label")
 }
